requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
boto3==1.42.88
botocore==1.42.88
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
limits==5.8.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
requests==2.33.1
resend==2.27.0
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
sse-starlette==3.3.4
standardwebhooks==1.0.1
starlette==1.0.0
svix==1.90.0
types-Deprecated==1.3.1.20260408
types-python-dateutil==2.9.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
wrapt==2.1.2
